+++ tests/AlternateCodec.php
@@ -27,7 +27,7 @@ class AlternateCodec extends JsonCodec {
 				}
 
 				/** @inheritDoc */
-				public function jsonClassHintFor( string $className, string $keyName ): ?string {
+				public function jsonClassHintFor( string $className, string $keyName ) {
 					return null;
 				}
 			};